fix(workflows): settle partial viewport-row evidence
The prior layer commits remain intact by instruction; this tip supersedes their conflicting completion wording and records the cumulative #2223 state as partial. The graph overlay frame and stage-chat line count still read host terminal.rows because PiCustomComponent.render supplies only a width, and VIEW_LINE_COUNT=32 remains the unhosted stage-chat fallback.

Narrow GraphView's host surface to the terminal.rows slice while retaining the full TUI for StageChatView's editor and chat-session host. Share the typed makeTestTui fixture across unit and manual callers, remove fixture casts and bare navigation literals, and restore tab-consumption coverage in the structured editor prompt test.

Supplying piTui constructs the production Editor and flips the long-prompt test's prompt-editor guards, so wheel events replace home/end there; keyboard coverage remains in the archive and prompt-height tests.

Refs: #2223
Assistant-model: GPT-5.6 Luna
