+++ packages/workflows/src/tui/graph-view-types.ts
@@ -4,6 +4,13 @@ import type { Store } from "../shared/store.js";
 import type { GraphTheme } from "./graph-theme.js";
 
 export type GraphViewMode = "overlay" | "widget";
+/**
+ * Terminal geometry exposed to GraphView. The renderer reads only `rows`;
+ * the attach shell may pass its full host TUI because it satisfies this slice.
+ */
+export type GraphViewHost = {
+	readonly terminal: Pick<TUI["terminal"], "rows">;
+};
 
 export interface GraphViewOpts {
 	mode: GraphViewMode;
@@ -43,12 +50,12 @@ export interface GraphViewOpts {
 	initialFocusedStageId?: string;
 	initialFocusedRunId?: string;
 	/**
-	 * Host TUI used for live terminal geometry. The overlay adapter passes the
-	 * same concrete host object to the graph and attached stage chat; GraphView
-	 * reads its terminal rows while the custom overlay renders and keeps the
-	 * hosted frame stable during teardown.
+	 * Host terminal geometry for the fullscreen frame. GraphView consumes only
+	 * `terminal.rows`; the attach shell passes the same host object to GraphView
+	 * and the attached stage chat. StageChatView retains the full TUI because its
+	 * editor and chat session host use it as well.
 	 */
-	piTui?: TUI;
+	piTui?: GraphViewHost;
 	/**
 	 * Invoked on each animation tick (~10 FPS) so the host can call
 	 * `tui.requestRender()`. Only wired in `overlay` mode; supplying it
