+++ test/unit/overlay-graph-helpers.ts
@@ -1,7 +1,6 @@
 /** Shared fixtures for overlay graph tests. */
 
 import assert from "node:assert/strict";
-import type { TUI } from "@earendil-works/pi-tui";
 import type { Store } from "../../packages/workflows/src/shared/store.js";
 import type {
 	PendingPrompt,
@@ -149,18 +148,7 @@ export function makeStore(snap: StoreSnapshot): Store {
 export const defaultTheme = deriveGraphTheme({});
 export const ANSI_RE = /\x1b\[[0-9;]*m/g;
 
-export function makeTestTui(rows: number | (() => number | undefined)): TUI {
-	const readRows = typeof rows === "function" ? rows : () => rows;
-	return {
-		requestRender: () => {},
-		terminal: {
-			get rows() {
-				return readRows();
-			},
-			columns: 80,
-		},
-	} as unknown as TUI;
-}
+export { makeTestTui } from "../support/fake-tui.js";
 export const SGR_MOUSE_WHEEL_DOWN = "\x1b[<65;10;10M";
 
 export function visibleText(lines: string[]): string {
