+++ test/unit/overlay-graph-navigation-03.test.ts
@@ -23,6 +23,7 @@ const {
 	makeInputRequest,
 	makeStore,
 	makeRun,
+	makeTestTui,
 	defaultTheme,
 	SGR_MOUSE_WHEEL_DOWN,
 	visibleText,
@@ -69,7 +70,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onStageAttach: (runId, stageId) => attached.push({ runId, stageId }),
 		});
 
@@ -90,7 +91,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onStageAttach: (_runId, stageId) => attached.push(stageId),
 		});
 
@@ -112,7 +113,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onStageAttach: (_runId, stageId) => attached.push(stageId),
 		});
 
@@ -143,7 +144,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onStageAttach: (_runId, stageId) => attached.push(stageId),
 		});
 
@@ -178,7 +179,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onStageAttach: (_runId, stageId) => attached.push(stageId),
 		});
 
@@ -231,7 +232,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 		});
 
 		view.render(96);
@@ -258,7 +259,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 			onDetach: () => {
 				detached += 1;
 			},
@@ -350,7 +351,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 		});
 
 		assert.doesNotMatch(visibleText(view.render(96)), /stage-5/);
@@ -375,7 +376,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 32 } },
+			piTui: makeTestTui(32),
 		});
 
 		view.render(96);
@@ -396,7 +397,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store: makeStore(makeSnap(stages)),
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 16 } },
+			piTui: makeTestTui(16),
 		});
 
 		view.render(96);
@@ -420,7 +421,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store: makeStore(makeSnap(stages)),
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 16 } },
+			piTui: makeTestTui(16),
 		});
 
 		view.render(96);
@@ -452,7 +453,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store: makeStore(makeRunPromptSnap(stages, makePendingPrompt({ message: "Continue?" }))),
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 16 } },
+			piTui: makeTestTui(16),
 		});
 
 		const lines = view.render(96);
@@ -472,7 +473,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 16 } },
+			piTui: makeTestTui(16),
 		});
 		const width = 80;
 		const message = "waiting for stage events…";
@@ -496,7 +497,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: null,
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 42 } },
+			piTui: makeTestTui(42),
 		});
 		const lines = view.render(96);
 		assert.equal(lines.length, 42);
@@ -556,7 +557,7 @@ describe("GraphView keyboard navigation", () => {
 			runId: "run-1",
 			store,
 			graphTheme: defaultTheme,
-			piTui: { terminal: { rows: 20 } },
+			piTui: makeTestTui(20),
 		});
 		const lines = view.render(40);
 		assert.equal(lines.length, 20);
