+++ test/unit/stage-chat-view-01.test.ts
@@ -12,7 +12,6 @@ import {
 	StageChatView,
 	setupRun,
 	stripAnsi,
-	type TUI,
 } from "./stage-chat-view-helpers.js";
 
 describe("StageChatView", () => {
@@ -288,10 +287,7 @@ describe("StageChatView", () => {
 			handle,
 			onDetach: () => {},
 			onClose: () => {},
-			piTui: {
-				requestRender: () => {},
-				terminal: { rows: 32, columns: 80 },
-			} as unknown as TUI,
+			piTui: makeTestTui(32),
 			piTheme: {},
 			piKeybindings: makeFakeKeybindings(),
 			piEditorFactory: () => {
@@ -358,7 +354,7 @@ describe("StageChatView", () => {
 		view.dispose();
 	});
 
-	test("structured editor prompt navigation, scroll, and Escape stay pending", async () => {
+	test("structured editor prompt navigation, tab, scroll, and Escape stay pending", async () => {
 		const store = createStore();
 		setupRun(store, "run-1", "stage-a");
 		const prompt = makePendingPrompt({
@@ -392,7 +388,7 @@ describe("StageChatView", () => {
 		});
 
 		view.render(80);
-		for (const key of ["\x1b[A", "\x1b[B", "\x1b[5~", "\x1b[6~", "\x1b"]) {
+		for (const key of ["\x1b[A", "\x1b[B", "\x1b[5~", "\x1b[6~", "\t", "\x1b"]) {
 			assert.equal(view.handleInput(key), true, JSON.stringify(key));
 		}
 		await flush();
